revert obsolete, as %python_provide does it (undocumented)
kparal pointed out python2-libsolv was already obsoleting
python-libsolv, even though this wasn't written in the spec.
This puzzled us, but I found out today that the python_provide
macro actually also produces an Obsoletes: line, but this is
not documented anywhere besides the trac ticket:
https://fedorahosted.org/fpc/ticket/565
so we actually didn't need to add the Obsoletes: directly.
This reverts to the spec as of -3.